Bugfix/restore stuff

+++ app/src/main/java/net/kelmer/correostracker/ui/list/ParcelListFragment.kt
@@ -1,33 +1,35 @@
 package net.kelmer.correostracker.ui.list
 
-import android.os.Bundle
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.appcompat.widget.PopupMenu
-import android.view.Menu
-import android.view.MenuInflater
-import android.view.MenuItem
-import android.view.View
-import android.widget.Toast
-import kotlinx.android.synthetic.main.fragment_parcel_list.*
-import net.kelmer.correostracker.R
-import net.kelmer.correostracker.base.fragment.BaseFragment
-import net.kelmer.correostracker.data.Result
-import net.kelmer.correostracker.data.model.local.LocalParcelReference
-import net.kelmer.correostracker.ext.observe
-import net.kelmer.correostracker.ui.detail.DetailActivity
-import timber.log.Timber
 import android.content.ClipData
 import android.content.ClipboardManager
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.widget.TextView
-import androidx.appcompat.app.AlertDialog
-import androidx.core.content.ContextCompat
+import android.os.Bundle
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import android.view.View
+import android.widget.Toast
+import androidx.appcompat.widget.PopupMenu
+import androidx.appcompat.widget.SearchView
+import androidx.core.view.MenuItemCompat
+import androidx.recyclerview.widget.LinearLayoutManager
+import kotlinx.android.synthetic.main.fragment_parcel_list.empty_state
+import kotlinx.android.synthetic.main.fragment_parcel_list.rv_parcel_list
+import kotlinx.android.synthetic.main.fragment_parcel_list.swipe_refresh
+import net.kelmer.correostracker.R
+import net.kelmer.correostracker.base.fragment.BaseFragment
+import net.kelmer.correostracker.customviews.ConfirmDialog
+import net.kelmer.correostracker.data.Result
+import net.kelmer.correostracker.data.model.local.LocalParcelReference
 import net.kelmer.correostracker.ext.isVisible
+import net.kelmer.correostracker.ext.observe
+import net.kelmer.correostracker.ui.detail.DetailActivity
 import net.kelmer.correostracker.ui.featuredialog.featureBlurbDialog
 import net.kelmer.correostracker.ui.list.adapter.ParcelClickListener
 import net.kelmer.correostracker.ui.list.adapter.ParcelListAdapter
+import timber.log.Timber
 
 
 /**
@@ -69,7 +71,11 @@ class ParcelListFragment : BaseFragment<ParcelListViewModel>() {
                     //do your things in each of the following cases
                     when (item.itemId) {
                         R.id.menu_delete -> {
-                            viewModel.deleteParcel(parcelReference)
+                            ConfirmDialog.confirmDialog(requireContext(),
+                                    R.string.delete_confirm_title,
+                                    R.string.delete_confirm_desc){
+                                viewModel.deleteParcel(parcelReference)
+                            }
                             true
                         }
                         R.id.menu_enable_notifications -> {
@@ -111,6 +117,7 @@ class ParcelListFragment : BaseFragment<ParcelListViewModel>() {
                 is Result.Success -> {
                     empty_state.isVisible = it.data.isEmpty()
                     adapter.updateItems(it.data)
+                    adapter.filter(searchView?.query?.toString() ?: "")
                 }
                 is Result.Failure -> {
                     Toast.makeText(context, "ERROR!!!", Toast.LENGTH_LONG).show()
@@ -152,8 +159,8 @@ class ParcelListFragment : BaseFragment<ParcelListViewModel>() {
     }
 
     private fun refreshFromRemote() {
-        viewModel.refresh(adapter.items)
-        adapter.items.forEach { p ->
+        viewModel.refresh(adapter.getAllItems())
+        adapter.getAllItems().forEach { p ->
             adapter.setLoading(p.code, true)
         }
     }
@@ -163,17 +170,32 @@ class ParcelListFragment : BaseFragment<ParcelListViewModel>() {
         viewModel.retrieveParcelList()
     }
 
+    var searchView: SearchView? = null
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         inflater.inflate(R.menu.main, menu)
 
+        val searchViewItem = menu.findItem(R.id.app_search)
+        searchView = MenuItemCompat.getActionView(searchViewItem) as SearchView
+        searchView?.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
+            override fun onQueryTextSubmit(query: String): Boolean {
+                searchView?.clearFocus()
+                adapter.filter(query)
+                return false
+            }
+
+            override fun onQueryTextChange(newText: String): Boolean {
+                adapter.filter(newText)
+                return false
+            }
+        })
         super.onCreateOptionsMenu(menu, inflater);
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             R.id.app_refresh_all -> {
-                viewModel.refresh(adapter.items)
-                adapter.items.forEachIndexed { i, p ->
+                viewModel.refresh(adapter.getAllItems())
+                adapter.getAllItems().forEachIndexed { i, p ->
                     adapter.setLoading(p.code, true)
                 }
             }
